Fix some formatting quirks and improve layout.
Newlines are counted as spaces and thus some of the layout was formatted
with uneven margins. Particularly the links in the top right and the
information above the pastes. I fixed that by using unordered lists.
Furthermore, Chris Granger had some tips for laying things out better.
They included adding a horizontal margin of 50px and a little padding
on the left side of the line numbers.